+++ include/nbl/video/utilities/CSubpassKiln.h
@@ -1,0 +1,25 @@
+// Copyright (C) 2018-2020 - DevSH Graphics Programming Sp. z O.O.
+// This file is part of the "Nabla Engine".
+// For conditions of distribution and use, see copyright notice in nabla.h
+
+#ifndef _NBL_VIDEO_C_SUBPASS_KILN_H_INCLUDED_
+#define _NBL_VIDEO_C_SUBPASS_KILN_H_INCLUDED_
+
+
+#include "nbl/video/utilities/ISubpassKiln.h"
+
+
+namespace nbl::video
+{
+    
+class CSubpassKiln final : public ISubpassKiln
+{
+    public:
+
+    protected:
+};
+
+
+}
+
+#endif
